Separate Management component better
